requirements.txt
playwright==1.52.0
psycopg2-binary==2.9.10
openai==1.82.0
python-dotenv==1.1.0
fastapi==0.115.0
uvicorn==0.34.0
